Fix MicroProfileDumpFileImmediately to also reset counters and clean up MicroProfileDumpFile
In change to remove duplicate file writing, the first file write was chosen. The second happens to also reset the counters for microprofile. Update the code to defer the write to the one that happens during MicroProfileFlip.

MicroProfileDumpFile did not synchronize variables properly. Clean this up and also defer it's file write to MicroProfileFip (by removing redundant first write).