Fix MicroProfileDumpFileImmediately to also reset counters and clean up MicroProfileDumpFile (#6)
Fix MicroProfileDumpFileImmediately to also reset counters and clean up MicroProfileDumpFile

MicroProfileDumpFileImmediately would write to the log immediately and, via global state, write a second time when MicroProfileFlip is next called. Consolidate the cleanup of the global state so that after the log file is written the global state is cleaned up. This results in only one write instead of two.

MicroProfileDumpFile did not synchronize variables properly when updating global state.